+++ app/src/main/java/com/mobilestudio/developer/features/home/presentation/viewmodel/HomeViewModel.kt
@@ -19,7 +19,7 @@ class HomeViewModel(
 
     fun getThemes() = themes
 
-    init {
+    fun init() {
         loadThemes()
     }
