+++ app/src/main/res/layout/fragment_theme_detail.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.mobilestudio.developer.features.themedetail.views.fragments.ThemeDetailFragment">
+        tools:context="com.mobilestudio.developer.features.themedetail.presentation.view.fragment.ThemeDetailFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
