+++ app/src/main/res/navigation/nav_graph.xml
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/subThemesFragment"
-        android:name="com.mobilestudio.developer.features.subthemes.views.fragments.SubThemesFragment"
+        android:name="com.mobilestudio.developer.features.subthemes.presentation.view.fragment.SubThemesFragment"
         android:label="fragment_sub_themes"
         tools:layout="@layout/fragment_sub_themes">
         <action
@@ -38,7 +38,7 @@
 
     <fragment
         android:id="@+id/themeDetailFragment"
-        android:name="com.mobilestudio.developer.features.themedetail.views.fragments.ThemeDetailFragment"
+        android:name="com.mobilestudio.developer.features.themedetail.presentation.view.fragment.ThemeDetailFragment"
         android:label="fragment_theme_detail"
         tools:layout="@layout/fragment_theme_detail">
         <action
@@ -55,7 +55,7 @@
 
     <fragment
         android:id="@+id/webViewerFragment"
-        android:name="com.mobilestudio.developer.features.webviewer.views.fragments.WebViewerFragment"
+        android:name="com.mobilestudio.developer.features.webviewer.presentation.view.fragment.WebViewerFragment"
         android:label="WebViewerFragment"
         tools:layout="@layout/fragment_web_viewer" >
         <argument
